Rename auth to login
